requirements.txt
# Core API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP and file handling
httpx==0.25.2
aiofiles==23.2.1
requests==2.31.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Machine learning (lightweight)
sentence-transformers==2.2.2
numpy==1.24.3

# Optional: Qdrant client (only if you want to use Qdrant later)
# qdrant-client==1.7.0

# Optional: OpenAI (alternative to Grok)
# openai==1.3.7

# Utilities
tiktoken==0.5.2